+++ tests/test_pipe+masterworker.cpp
@@ -79,7 +79,7 @@ public:
         ++(*t), ++numtasks;
         return task;
     }
-    void eosnotify(int id) {
+    void eosnotify(ssize_t id) {
         if (id == -1) {
             eosarrived= true;
             if (numtasks==0) lb->broadcast_task(EOS);
